--- packages/avatar/react.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/react/index.js')
