+++ packages/avatar/src/css/index.ts
@@ -4,7 +4,7 @@ import {
   colorsBackgroundUtility
 } from '@pluralsight/ps-design-system-core'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   '.psds-avatar': {
