+++ packages/avatar/src/js/index.ts
@@ -1,6 +1,6 @@
 import sha1 from 'tiny-hashes/sha1'
 
-import { colors, fallbackPixel } from '../vars'
+import { colors, fallbackPixel } from '../vars/index.js'
 
 export const getInitials = (fullname?: string) => {
   if (!fullname || !fullname.length) return ':)'
